Worked on merging recursive hasher examples and volume collector #107
